Revert "Merge branch 'main' into origin/NguyenTCP"
This reverts commit 0bb914578187dda9ab8bf86392ee755558ab5e2e, reversing
changes made to 76a420316dd4791fb209041fb7442567685b0da9.

+++ Backend/Warehouse.Entities/Models/Brand.cs
@@ -5,7 +5,7 @@ namespace Warehouse.Entities.Models;
 
 public partial class Brand
 {
-    public long BrandId { get; set; }
+    public int BrandId { get; set; }
 
     public string BrandName { get; set; } = null!;
